Move inferLanguage to its own module
